requirements.txt
neo4j>=5.0.0
streamlit==1.29.0
python-dotenv>=0.19.0
pandas==2.1.3
watchdog 